rename: CustomInput input 폴더 아래 위치로 수정
